09_Docker/Projects/01_PYTHON_CRUD_APPLICATION/requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
cryptography